+++ packages/@headlessng-examples/src/lib/atoms/index.ts
@@ -1,4 +1,5 @@
 export * from './checkbox';
 export * from './description';
+export * from './error-message';
 export * from './field';
 export * from './label';