+++ packages/@headlessng-primitives/error-message/src/lib/error-message.directive.ts
@@ -1,0 +1,16 @@
+import { Directive, OnDestroy } from '@angular/core';
+import { ErrorMessageFieldRef } from '@headlessng/primitives/field';
+
+@Directive({
+  exportAs: 'hErrorMessageRef',
+  host: {
+    '[attr.id]': 'id()'
+  },
+  selector: '[hErrorMessage]',
+  standalone: true
+})
+export class ErrorMessageDirective extends ErrorMessageFieldRef implements OnDestroy {
+  public ngOnDestroy(): void {
+    this.destroyRef();
+  }
+}